Added the ability to add the entire armorset to a wishlist

+++ app/src/main/java/com/ghstudios/android/data/classes/meta/ArmorMetadata.kt
@@ -7,7 +7,7 @@ data class ArmorMetadata(
     val name: String,
     val slot: String,
     val rarity: Int,
-    val family: Int,
+    val family: Long,
     val familyName: String,
     val icon_name: String
 ) : ITintedIcon {
